refactor: app 폴더를 src 밑으로 이동
