+++ src/components/home/DisplayProducts.tsx
