+++ src/components/home/ProductInfo.tsx
